Moved all of the file handling functions to their own mod.
